requirements.txt
aniso8601==8.0.0
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-RESTful==0.3.7
itsdangerous==1.1.0
lxml==4.4.1
MarkupSafe==1.1.1
pymongo==3.6.0
python-docx==0.8.10
python-tr==0.1.2
pytz==2019.3
six==1.13.0
waitress==1.3.1
Werkzeug==0.16.0
